Custom Star Menu Folder with Office 2016 Shortcuts
Some users on Windows 7 still want Office application shortcuts grouped
in one start menu folder.  Office 2016 is meant for Windows 10 and the
"tiled format", therefore when installing on Windows 7, the shortcuts
are scattered on the main program bar.

This script simply created a new folder, loops through all of the
shortcuts and copies the shortcuts with the wild card express *2016* in
the name to the newly created programs folder.